weather_mv/loader_pipeline/requirements.txt
apache-beam[gcp]==2.40.0
dataclasses==0.6
numpy==1.22.4
pandas==1.4.3
xarray==2022.3.0
google-cloud-storage[bqstorage, pandas]==2.2.1
cfgrib==0.9.10.1
cloudpickle==2.1.0
rioxarray==0.11.1
netCDF4==1.6.0
geojson==2.5.0
more-itertools==8.13.0
simplejson==3.17.6
shapely==1.8.2
rasterio==1.3.0
earthengine-api==0.1.317
pyproj==3.3.1
gdal[numpy]==3.5.1
google-api-python-client==1.12.11
